Fixing core consultarPlanes and consultarPlan and change int to bigint in planes
Synch PlanBean with planes table

+++ project/workspace/concesionarias/db/src/main/resources/liquibase/tables/cuotas.sql
@@ -1,6 +1,6 @@
 CREATE TABLE cuotas (
-	id_plan INT FOREIGN KEY REFERENCES planes(id)
-	, id_cuota INT NOT NULL
+	id_plan BIGINT FOREIGN KEY REFERENCES planes(id)
+	, id_cuota BIGINT NOT NULL
 	, fecha_pago DATETIME NULL
 	, fecha_vencimiento DATETIME NOT NULL
 	, PRIMARY KEY (id_plan, id_cuota)
